Replace type Client with map to represent base-url and telegram token
The goal to create new type Client was to provide base-url and telegram token to each
api functions. Even though it is achieved, each time to invoke api
functions, object client needs to be passed. And type Client actually
can be replaced with map. And map has benefit that it doesn't hide
data so it can be processed by other functions.

This commit creates telegram-api-test file and move couples of tests
from bot_test to just created file.